# docs/swagger.yaml
basePath: /api
definitions:
  request.ParameterListRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
  request.ParameterRequest:
    properties:
      description:
        maxLength: 255
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - description
    - name
    type: object
  response.ParameterListResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  response.ParameterResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  response.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: This is a siakad api service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SIAKAD API
  version: "1.0"
paths:
  /parameter-lists:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ParameterListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ParameterListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      tags:
      - Parameter List
  /parameter-lists/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search
        in: query
        name: search
        type: string
      - description: Parameter ID
        in: query
        name: parameterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.ParameterListResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      tags:
      - Parameter List
  /parameter-lists/parameter-list/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ParameterListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      tags:
      - Parameter List
    put:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ParameterListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ParameterListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      tags:
      - Parameter List
  /parameters:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ParameterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ParameterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      tags:
      - Parameter
  /parameters/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.ParameterResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      tags:
      - Parameter
  /parameters/parameter/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ParameterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      tags:
      - Parameter
    put:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ParameterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ParameterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      tags:
      - Parameter
swagger: "2.0"